refactor(): Override all methods from WebIdlGrammarDefinition
In WebIdlParserDefinition organize all methods to match the ordering of WebIdlGrammarDefinition. Add an override that just calls the super interface for all missing methods.

There is a lint that recommends not having an override that simple calls the super method. Temporarily turn this off until all of the super interface is properly overriden.